requirements.txt
fastapi~=0.104.1
uvicorn~=0.23.2
sqlalchemy~=2.0.21
psycopg2-binary~=2.9.9
python-dotenv~=1.0.0
passlib[bcrypt]~=1.7.4              # Şifre hashleme için
python-jose[cryptography]~=3.3.0    # JWT token işlemleri için
python-multipart~=0.0.6             # Dosya yükleme (UploadFile) için
boto3~=1.34.116                     # MinIO (S3 uyumlu) depolama ile etkileşim için